--- src/main/kotlin/topKFrequentElements/TopKFrequentElements.kt
@@ -1,20 +1,0 @@
-package topKFrequentElements
-
-class TopKFrequentElements {
-    fun topKFrequent(nums: IntArray, k: Int): IntArray {
-        val freqPairList = getFreqPairList(nums)
-        return if (k >= freqPairList.size) {
-            freqPairList.map { (key, _) -> key }.toIntArray()
-        } else {
-            freqPairList.subList(0, k).map { (key, _) -> key }.toIntArray()
-        }
-    }
-
-    private fun getFreqPairList(nums: IntArray): List<Pair<Int, Int>> {
-        val result = mutableMapOf<Int, Int>()
-        nums.forEach { num ->
-            result[num] = result.getOrDefault(num, 0) + 1
-        }
-        return result.toList().sortedByDescending { (_, value) -> value }
-    }
-}
